chore: move ref to hook
